+++ backend/src/main/java/org/example/controller/TaskController.java
@@ -3,6 +3,7 @@ package org.example.controller;
 import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
+import java.util.Objects;
 import org.example.MyConfiguration;
 import org.example.model.KeywordsForTasks;
 import org.example.model.Task;
@@ -12,6 +13,7 @@ import org.example.service.KeywordsForTasksService;
 import org.example.service.TaskService;
 import org.example.service.UserService;
 import org.example.utils.exceptions.ServiceException;
+import org.example.utils.exceptions.ServiceExceptionType;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Import;
 import org.springframework.http.HttpStatus;
@@ -38,11 +40,26 @@ public class TaskController
 
 	@RequestMapping(value = "/", method = RequestMethod.POST)
 	@ResponseStatus(HttpStatus.CREATED)
-	public Task addTask(@RequestBody Task task)
+	public Task addTask(@RequestBody Task task,
+		@RequestHeader(name = "Authorization") String token)
 	{
 		try
 		{
-			//TODO: Validate ownerID with token
+			User senderByToken = userService.getUserByToken(token.substring(7));
+			User senderByObject = null;
+
+			if (task.getOwnerid() != null)
+			{
+				senderByObject = userService.getUserById(task.getOwnerid());
+			}
+
+			if (senderByObject != null && senderByToken != null && !Objects.equals(senderByToken.getId(), senderByObject.getId()))
+			{
+				throw new ServiceException(ServiceExceptionType.UNAUTHORIZED_TO_CHANGE,
+					"Creating tasks in the name of other's are not allowed!"
+				);
+			}
+
 			return taskService.saveTask(task);
 		}
 		catch (ServiceException e)
@@ -67,10 +84,27 @@ public class TaskController
 
 	@RequestMapping(value = "/", method = RequestMethod.PUT)
 	@ResponseStatus(HttpStatus.OK)
-	public Task updateTask(@RequestBody Task task)
+	public Task updateTask(@RequestBody Task task,
+		@RequestHeader(name = "Authorization") String token)
 	{
 		try
 		{
+			User requestedUser = userService.getUserByToken(token.substring(7));
+
+			Task taskInDatabase = null;
+			if (task.getId() != null)
+			{
+				taskInDatabase = taskService.getTaskById(task.getId());
+			}
+
+			if (taskInDatabase != null
+				&& !Objects.equals(requestedUser.getId(), taskInDatabase.getOwnerid()))
+			{
+				throw new ServiceException(ServiceExceptionType.UNAUTHORIZED_TO_CHANGE,
+					"Changing others task data is not allowed!"
+				);
+			}
+
 			return taskService.updateTask(task);
 		}
 		catch (ServiceException e)
@@ -79,6 +113,9 @@ public class TaskController
 			{
 				case ID_NOT_GIVEN, INVALID_ARGUMENT:
 					throw new ResponseStatusException(HttpStatus.BAD_REQUEST, e.getMessage());
+
+				case UNAUTHORIZED_TO_CHANGE:
+					throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, e.getMessage());
 
 				case CONSTRAINT_VIOLATION, ID_NOT_FOUND:
 					throw new ResponseStatusException(HttpStatus.NOT_FOUND, e.getMessage());
@@ -95,10 +132,20 @@ public class TaskController
 
 	@RequestMapping(value = "/maintaskid", method = RequestMethod.PUT)
 	@ResponseStatus(HttpStatus.OK)
-	public Integer updateMainTaskIdForTask(@RequestBody Task task)
+	public Integer updateMainTaskIdForTask(@RequestBody Task task,
+		@RequestHeader(name = "Authorization") String token)
 	{
 		try
 		{
+			User requestedUser = userService.getUserByToken(token.substring(7));
+
+			if (!Objects.equals(requestedUser.getId(), task.getOwnerid()))
+			{
+				throw new ServiceException(ServiceExceptionType.UNAUTHORIZED_TO_CHANGE,
+					"Changing others task data is not allowed!"
+				);
+			}
+
 			return taskService.setMainTaskId(task.getId(), task.getMaintaskid());
 		}
 		catch (ServiceException e)
@@ -107,6 +154,9 @@ public class TaskController
 			{
 				case ID_NOT_GIVEN, NULL_ARGUMENT, INVALID_ARGUMENT:
 					throw new ResponseStatusException(HttpStatus.BAD_REQUEST, e.getMessage());
+
+				case UNAUTHORIZED_TO_CHANGE:
+					throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, e.getMessage());
 
 				case CONSTRAINT_VIOLATION, ID_NOT_FOUND:
 					throw new ResponseStatusException(HttpStatus.NOT_FOUND, e.getMessage());
@@ -121,7 +171,6 @@ public class TaskController
 		}
 	}
 
-	//TODO: Write test for not completed tasks
 	@RequestMapping(value = "/", method = RequestMethod.GET)
 	@ResponseStatus(HttpStatus.OK)
 	public Iterable<TaskListingFilter> listTasksByFilter(@RequestParam(required = false, value = "keywords") Collection<String> keywords,
@@ -212,10 +261,25 @@ public class TaskController
 
 	@RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
 	@ResponseStatus(HttpStatus.NO_CONTENT)
-	public void deleteTask(@PathVariable(value = "id") Long taskId)
+	public void deleteTask(@PathVariable(value = "id") Long taskId,
+		@RequestHeader(name = "Authorization") String token)
 	{
 		try
 		{
+			User requestedUser = userService.getUserByToken(token.substring(7));
+			Task task = null;
+			if (taskId != null)
+			{
+				task = taskService.getTaskById(taskId);
+			}
+
+			if (task != null && !Objects.equals(requestedUser.getId(), task.getOwnerid()))
+			{
+				throw new ServiceException(ServiceExceptionType.UNAUTHORIZED_TO_CHANGE,
+					"Changing others task data is not allowed!"
+				);
+			}
+
 			taskService.deleteTask(taskId);
 		}
 		catch (ServiceException e)
@@ -224,6 +288,8 @@ public class TaskController
 			{
 				case ID_NOT_FOUND:
 					throw new ResponseStatusException(HttpStatus.NOT_FOUND, e.getMessage());
+				case UNAUTHORIZED_TO_CHANGE:
+					throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, e.getMessage());
 				default:
 					throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, e.getMessage());
 			}
@@ -236,11 +302,37 @@ public class TaskController
 
 	@RequestMapping(value = "/keyword/", method = RequestMethod.POST)
 	@ResponseStatus(HttpStatus.CREATED)
-	public Iterable<KeywordsForTasks> addKeyword(@RequestBody Iterable<KeywordsForTasks> keyword)
+	public Iterable<KeywordsForTasks> addKeyword(@RequestBody Iterable<KeywordsForTasks> keywords,
+		@RequestHeader(name = "Authorization") String token)
 	{
 		try
 		{
-			return keywordsForTasksService.saveKeywordsForTasks(keyword);
+			User requestedUser = userService.getUserByToken(token.substring(7));
+
+			boolean isTaskIdsValid = true;
+
+			for (var keyword : keywords)
+			{
+				if (keyword.getTaskid() == null || taskService.getTaskById(keyword.getTaskid()) == null)
+				{
+					isTaskIdsValid = false;
+				}
+			}
+
+			Task task = null;
+			if (isTaskIdsValid)
+			{
+				task = taskService.getTaskById(keywords.iterator().next().getTaskid());
+			}
+
+			if (task != null && !Objects.equals(requestedUser.getId(), task.getOwnerid()))
+			{
+				throw new ServiceException(ServiceExceptionType.UNAUTHORIZED_TO_CHANGE,
+					"Changing others task's keyword data is not allowed!"
+				);
+			}
+
+			return keywordsForTasksService.saveKeywordsForTasks(keywords);
 		}
 		catch (ServiceException e)
 		{
@@ -248,6 +340,9 @@ public class TaskController
 			{
 				case NULL_ARGUMENT, ID_GIVEN, INVALID_ARGUMENT:
 					throw new ResponseStatusException(HttpStatus.BAD_REQUEST, e.getMessage());
+
+				case UNAUTHORIZED_TO_CHANGE:
+					throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, e.getMessage());
 
 				case CONSTRAINT_VIOLATION:
 					throw new ResponseStatusException(HttpStatus.NOT_FOUND, e.getMessage());
@@ -278,16 +373,40 @@ public class TaskController
 
 	@RequestMapping(value = "/keyword/{id}", method = RequestMethod.DELETE)
 	@ResponseStatus(HttpStatus.NO_CONTENT)
-	public void deleteKeyword(@PathVariable(value = "id") Long id)
+	public void deleteKeyword(@PathVariable(value = "id") Long id,
+		@RequestHeader(name = "Authorization") String token)
 	{
 		try
 		{
+			User requestedUser = userService.getUserByToken(token.substring(7));
+
+			KeywordsForTasks keyword = null;
+			if (id != null)
+			{
+				keyword = keywordsForTasksService.getKeywordForTaskById(id);
+			}
+
+			Task task = null;
+			if (keyword != null)
+			{
+				task = taskService.getTaskById(keyword.getTaskid());
+			}
+
+			if (task != null && !Objects.equals(requestedUser.getId(), task.getOwnerid()))
+			{
+				throw new ServiceException(ServiceExceptionType.UNAUTHORIZED_TO_CHANGE,
+					"Changing others task's keyword data is not allowed!"
+				);
+			}
+
 			keywordsForTasksService.deleteKeywordForTask(id);
 		}
 		catch (ServiceException e)
 		{
 			switch (e.getServiceExceptionTypeEnum())
 			{
+				case UNAUTHORIZED_TO_CHANGE:
+					throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, e.getMessage());
 				case ID_NOT_FOUND:
 					throw new ResponseStatusException(HttpStatus.NOT_FOUND, e.getMessage());
 				default:
@@ -302,16 +421,29 @@ public class TaskController
 
 	@RequestMapping(value = "{id}/keyword", method = RequestMethod.DELETE)
 	@ResponseStatus(HttpStatus.NO_CONTENT)
-	public void deleteKeywordsByTaskid(@PathVariable(value = "id") Long taskId)
+	public void deleteKeywordsByTaskid(@PathVariable(value = "id") Long taskId,
+		@RequestHeader(name = "Authorization") String token)
 	{
 		try
 		{
+			User requestedUser = userService.getUserByToken(token.substring(7));
+			Task task = taskService.getTaskById(taskId);
+
+			if (!Objects.equals(requestedUser.getId(), task.getOwnerid()))
+			{
+				throw new ServiceException(ServiceExceptionType.UNAUTHORIZED_TO_CHANGE,
+					"Changing others task's keyword data is not allowed!"
+				);
+			}
+
 			keywordsForTasksService.deleteKeywordsByTaskId(taskId);
 		}
 		catch (ServiceException e)
 		{
 			switch (e.getServiceExceptionTypeEnum())
 			{
+				case UNAUTHORIZED_TO_CHANGE:
+					throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, e.getMessage());
 				case ID_NOT_FOUND:
 					throw new ResponseStatusException(HttpStatus.NOT_FOUND, e.getMessage());
 				default:
